fix: uncomment delay
